Simplified and abstracted code
